fix: move PrefetchBalancesWrapper
